Further decouple the stages and beef up tests
